bug-1812771: move Android_* processor rules to separate file
