+++ socorro/tests/processor/rules/test_general.py
@@ -237,25 +237,6 @@ class TestCPUInfoRule:
         rule.act(raw_crash, dumps, processed_crash, str(tmp_path), status)
         assert processed_crash["cpu_arch"] == "x86"
 
-    @pytest.mark.parametrize(
-        "android_cpu_abi, expected",
-        [
-            ("x86", "x86"),
-            ("arm64-v8a", "arm64"),
-            ("x86_64", "amd64"),
-            ("value not in map", "value not in map"),
-        ],
-    )
-    def test_cpu_arch_from_android_cpu_abi(self, tmp_path, android_cpu_abi, expected):
-        raw_crash = {"Android_CPU_ABI": android_cpu_abi}
-        processed_crash = {}
-        dumps = {}
-        status = Status()
-
-        rule = CPUInfoRule()
-        rule.act(raw_crash, dumps, processed_crash, str(tmp_path), status)
-        assert processed_crash["cpu_arch"] == expected
-
     def test_no_cpu_microcode_version(self, tmp_path):
         raw_crash = {}
         processed_crash = {}
